move EmailMapper in ./mappers
